// src/app/characters/interfaces/character.interface.ts
// Generated by https://quicktype.io

export interface Character {
  id:               string;
  character:        string;
  serie:        Serie;
  alter_ego:        string;
  live_actor: string;
  cohabitants:       string;
  alt_img?:         string;
}

export enum Serie {
  Aqnhqv = "Aqui no hay quien viva",
  Lqsa = "La que se avecina",
}